requirements.txt
openai>=1.0.0
pandas>=2.0.0
scipy>=1.11.0
python-dotenv>=1.0.0
tenacity>=8.2.0
# Milestone 7: RAG Pipeline
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-openai>=0.2.0
chromadb>=0.5.0
pypdf>=4.0.0
# Milestone 8: Multimodal
matplotlib>=3.8.0
numpy>=1.26.0
